完善createBid，并且将BidRank表删除bidId主键，采用bwicId + accountId作为联合主键，之前的重复创建的bug是因为从requestDTO modelMapper 到bid的时候会重复bidId，所以后面改为了手动设置。

+++ service/bid-api/src/main/java/com/ebidding/bid/api/BidCreateRequestDTO.java
@@ -10,8 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class BidCreateRequestDTO {
-    private int bwicId;
-
+    private Long bwicId;
     private Double price;
 }
